Remove `BaseTokenizerGroup`
